完善车辆信息展示 新增详情功能 上报车辆效果卡片展示 二次开发高德地图 新增行为id属性，用于为图片视频等媒体资源绑定行为(tob_key_picture表) 模拟拍摄车辆行为流程

+++ cigarette-system/src/main/java/com/ruoyi/system/service/ITobVehicleBehaviorService.java
@@ -31,7 +31,7 @@ public interface ITobVehicleBehaviorService
     /**
      * 查询车辆行为记录列表
      *
-     * @param voTobVehicleBehavior 车辆行为记录
+     * @param tobVehicleBehaviorVo 车辆行为记录
      * @return 车辆行为记录集合
      */
     public List<TobVehicleBehaviorVo> selectTobVehicleBehaviorVoList(TobVehicleBehaviorVo tobVehicleBehaviorVo );
@@ -42,7 +42,7 @@ public interface ITobVehicleBehaviorService
      * @param tobVehicleBehavior 车辆行为记录
      * @return 结果
      */
-    public int insertTobVehicleBehavior(TobVehicleBehavior tobVehicleBehavior);
+    public Long insertTobVehicleBehavior(TobVehicleBehavior tobVehicleBehavior);
 
     /**
      * 修改车辆行为记录
